+++ build-system/dep-check-config.js
@@ -291,6 +291,17 @@ exports.rules = [
     ],
   },
 
+  {
+    mustNotDependOn: [
+      /** DO NOT WHITELIST ANY FILES */
+      'ads/google/deprecated_doubleclick.js',
+      /** DO NOT WHITELIST ANY FILES */
+    ],
+    whitelist: [
+      '3p/integration.js->ads/google/deprecated_doubleclick.js',
+    ],
+  },
+
   // Delayed fetch for Doubleclick will be deprecated on March 29, 2018.
   // Doubleclick.js will be deleted from the repository at that time.
   // Please see https://github.com/ampproject/amphtml/issues/11834
@@ -314,7 +325,6 @@ exports.rules = [
       'ads/yieldbot.js->ads/google/doubleclick.js',
       /** DO NOT ADD TO WHITELIST **/
       'ads/criteo.js->ads/google/doubleclick.js',
-      '3p/integration.js->ads/google/doubleclick.js',
       /** DO NOT ADD TO WHITELIST **/
     ],
   },
